+++ src/main/java/net/sf/jabref/gui/maintable/MainTable.java
@@ -499,11 +499,14 @@ public class MainTable extends JTable {
     }
 
     public int findEntry(BibEntry entry) {
-        return model.getTableRows().indexOf(entry);
-    }
-
-    public int findLastEntry(BibEntry entry) {
-        return model.getTableRows().lastIndexOf(entry);
+        EventList<BibEntry> tableRows = model.getTableRows();
+        for (int row = 0; row < tableRows.size(); row++) {
+            BibEntry bibEntry = tableRows.get(row);
+            if (entry == bibEntry) { // NOPMD (equals doesn't recognise duplicates)
+                return row;
+            }
+        }
+        return -1;
     }
 
     /**